+++ stub/lang/en/captcha.php
@@ -1,0 +1,5 @@
+<?php
+return [
+    'enter_number' => "Enter verification number",
+    'wrong_number' => 'Incorrect verification number.',
+];