+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Minor helpers change
 - English captcha messages
 - Inline height CSS fix
+- Currency format locale fix
 
 #### [2.1.0] - 2025-07-27